I modularized. I might have messed up the functionality. BRB.
